Drop 'modern', not actually applicable
In order to be a useful for leaning modern engines it is also useful to
have samples to compare with legacy ones. Sometimes people who learn
a modern engine will need to do something in a legacy one.

Also 'modern' is entirely subjective and undefined. As such this
qualification was just glib jargon with no purpose in this context.